refactor(config_routes): extract system.py + config.py from _factory.py (2/8)
Schritt 2 von 8 im Split-Plan
(`docs/superpowers/plans/2026-04-29-config-routes-split.md`).

What moved
- `_register_system_routes()` (Zeilen 113-702 im Backbone-`_factory.py`,
  ~590 LOC) -> `cognithor.channels.config_routes.system._register_system_routes`.
  Registriert Dashboard, Status, Overview, Agents, Credentials, Bindings,
  Circles, Sandbox, Wizards, RBAC, Auth-Stats, Agent-Heartbeat.
- `_register_config_routes()` (Zeilen 705-1245, ~540 LOC) ->
  `cognithor.channels.config_routes.config._register_config_routes`.
  Registriert Health, Config-CRUD, Locales, Translation, Presets, Revisions,
  Network, Devices und das Config-Section-Routing.
- `_factory.py` ist von 6 628 auf 5 503 Zeilen geschrumpft; importiert die
  beiden Helper jetzt aus den Sub-Modulen und ruft sie unveraendert in der
  bestehenden Reihenfolge in `create_config_routes()` auf.

Per-Sub-Module-Imports
- `system.py`: `time`, `yaml`, `Request`, `HTTPException`, `Any`, `ConfigManager`.
- `config.py`: `Request`, `HTTPException`, `Any`, `ConfigManager`.
Keine Stdlib-Imports im Sub-Modul, die nicht tatsaechlich genutzt werden
(F401-clean). Lazy-Imports innerhalb der Handler bleiben verbatim, wie im
Plan vorgesehen.

Aufruf-Reihenfolge
`create_config_routes()` ruft `_register_system_routes()` weiter als ersten
und `_register_config_routes()` als zweiten Helper — exakt wie im
Backbone. FastAPI Path-Match-Reihenfolge unveraendert.

Public API
`from cognithor.channels.config_routes import create_config_routes` ist
weiter alleinige Public-Surface; keiner der 11 Konsumenten wird angefasst.

Verification
- `pytest tests/test_channels/test_config_routes.py::TestRouteRegistration::test_route_inventory_unchanged` -> passed (303/303 endpoints).
- Volle Konsumenten-Sweep ueber 10 Test-Files (`test_config_routes`, `test_workflow_graph_routes`, `test_knowledge_graph_routes`, `test_config_manager`, `test_ui_api_integration`, `test_f003`, `test_f005`, `test_f025`, `test_bootstrap_binding`, `test_prompt_evolution_api`, `test_wiring_all_four`) -> 396 passed.
- `ruff check src/cognithor/channels/config_routes/` -> All checks passed.
- `ruff format --check src/cognithor/channels/config_routes/` -> 4 files already formatted.
- `python -c "from cognithor.channels.config_routes import create_config_routes"` -> OK.
