+++ src/cognithor/channels/config_routes/_factory.py
@@ -51,6 +51,9 @@ if TYPE_CHECKING:
 log = get_logger(__name__)
 
 
+from cognithor.channels.config_routes.config import _register_config_routes
+from cognithor.channels.config_routes.system import _register_system_routes
+
 # ======================================================================
 # Public entry-point
 # ======================================================================
@@ -108,1141 +111,6 @@ def create_config_routes(
     _register_autonomous_routes(app, deps, gateway)
     _register_feedback_routes(app, deps, gateway)
     _register_social_routes(app, deps, gateway)
-
-
-# ======================================================================
-# System / health / status / dashboard routes
-# ======================================================================
-
-
-def _register_system_routes(
-    app: Any,
-    deps: list[Any],
-    config_manager: ConfigManager,
-    gateway: Any,
-) -> None:
-    """Dashboard, status, overview, presets, bindings, agents, credentials."""
-
-    # -- Admin Dashboard --------------------------------------------------
-
-    @app.get("/dashboard")
-    async def serve_dashboard():
-        """Liefert das Admin-Dashboard als HTML."""
-        from pathlib import Path
-
-        dashboard_path = Path(__file__).parent.parent / "gateway" / "dashboard.html"
-        if dashboard_path.exists():
-            from starlette.responses import HTMLResponse
-
-            return HTMLResponse(dashboard_path.read_text(encoding="utf-8"))
-        return {"error": "Dashboard nicht gefunden"}
-
-    # -- Status -----------------------------------------------------------
-
-    @app.get("/api/v1/status", dependencies=deps)
-    async def get_system_status() -> dict[str, Any]:
-        """Gibt den aktuellen System-Status zurueck."""
-        status: dict[str, Any] = {
-            "timestamp": time.time(),
-            "config_version": config_manager.config.version,
-            "owner": config_manager.config.owner_name,
-        }
-
-        # RuntimeMonitor
-        try:
-            from cognithor.openclaw.runtime_monitor import RuntimeMonitor
-
-            monitor = RuntimeMonitor()
-            status["runtime"] = {"metrics_count": len(monitor._metrics)}
-        except Exception:
-            status["runtime"] = {"available": False}
-
-        # HeartbeatScheduler
-        try:
-            hb_config = config_manager.config.heartbeat
-            status["heartbeat"] = {
-                "enabled": hb_config.enabled,
-                "interval_minutes": hb_config.interval_minutes,
-                "channel": hb_config.channel,
-            }
-        except Exception:
-            status["heartbeat"] = {"available": False}
-
-        # Active Channels
-        ch = config_manager.config.channels
-        active_channels = []
-        for attr in dir(ch):
-            if attr.endswith("_enabled") and getattr(ch, attr, False):
-                active_channels.append(attr.replace("_enabled", ""))
-        status["active_channels"] = active_channels
-
-        # Models
-        models = config_manager.config.models
-        status["models"] = {
-            "planner": models.planner.name,
-            "executor": models.executor.name,
-            "coder": models.coder.name,
-            "embedding": models.embedding.name,
-        }
-
-        # LLM Backend
-        status["llm_backend"] = config_manager.config.llm_backend_type
-        return status
-
-    # -- Overview ---------------------------------------------------------
-
-    @app.get("/api/v1/overview", dependencies=deps)
-    async def get_overview() -> dict[str, Any]:
-        """Gibt eine kompakte Konfigurationsuebersicht zurueck."""
-        try:
-            from cognithor.gateway.config_api import ConfigManager as CfgMgr
-
-            cfg_mgr = CfgMgr(config_manager.config)
-            overview = cfg_mgr.get_overview()
-            return overview.model_dump()
-        except Exception:
-            log.exception("Failed to build configuration overview")
-            return {"error": "Konfigurationsübersicht konnte nicht geladen werden."}
-
-    # -- Agents -----------------------------------------------------------
-
-    @app.get("/api/v1/agents", dependencies=deps)
-    async def list_agents() -> dict[str, Any]:
-        """Listet alle registrierten Agent-Profile aus agents.yaml."""
-        try:
-            agents_path = config_manager.config.cognithor_home / "agents.yaml"
-            if agents_path.exists():
-                raw = yaml.safe_load(agents_path.read_text(encoding="utf-8")) or {}
-                agents = raw.get("agents", [])
-            else:
-                agents = [
-                    {
-                        "name": "jarvis",
-                        "display_name": "Jarvis",
-                        "description": "Haupt-Agent (Default)",
-                        "system_prompt": "",
-                        "language": "de",
-                        "trigger_patterns": [],
-                        "trigger_keywords": [],
-                        "priority": 100,
-                        "allowed_tools": [],
-                        "blocked_tools": [],
-                        "preferred_model": "",
-                        "temperature": 0.7,
-                        "top_p": None,
-                        "enabled": True,
-                    }
-                ]
-            return {"agents": agents}
-        except Exception as exc:
-            log.error("agents_list_failed", error=str(exc))
-            return {"agents": [], "error": "Agenten konnten nicht geladen werden"}
-
-    @app.get("/api/v1/agents/{agent_name}", dependencies=deps)
-    async def get_agent(agent_name: str) -> dict[str, Any]:
-        """Get a single agent by name."""
-        # Try agents.yaml first
-        agents_path = config_manager.config.cognithor_home / "agents.yaml"
-        if agents_path.exists():
-            raw = yaml.safe_load(agents_path.read_text(encoding="utf-8")) or {}
-            for a in raw.get("agents", []):
-                if a.get("name") == agent_name:
-                    return a
-
-        # Fallback: check the live agent router
-        router = getattr(gateway, "_agent_router", None) if gateway else None
-        if router:
-            agent_obj = getattr(router, "get_agent", lambda n: None)(agent_name)
-            if agent_obj:
-                return {
-                    "name": getattr(agent_obj, "name", agent_name),
-                    "display_name": getattr(agent_obj, "display_name", agent_name.title()),
-                    "description": getattr(agent_obj, "description", ""),
-                    "system_prompt": getattr(agent_obj, "system_prompt", ""),
-                    "language": getattr(agent_obj, "language", "de"),
-                    "preferred_model": getattr(agent_obj, "preferred_model", ""),
-                    "temperature": getattr(agent_obj, "temperature", 0.7),
-                    "top_p": getattr(agent_obj, "top_p", None),
-                    "priority": getattr(agent_obj, "priority", 0),
-                    "enabled": getattr(agent_obj, "enabled", True),
-                    "allowed_tools": getattr(agent_obj, "allowed_tools", None) or [],
-                    "blocked_tools": getattr(agent_obj, "blocked_tools", []),
-                    "can_delegate_to": getattr(agent_obj, "can_delegate_to", []),
-                    "sandbox_timeout": getattr(agent_obj, "sandbox_timeout", 30),
-                    "sandbox_network": getattr(agent_obj, "sandbox_network", "allow"),
-                }
-
-        # Last fallback for default "jarvis"
-        if agent_name == "jarvis":
-            return {
-                "name": "jarvis",
-                "display_name": "Jarvis",
-                "description": "Default Agent",
-                "system_prompt": "",
-                "language": "de",
-                "preferred_model": "",
-                "temperature": 0.7,
-                "top_p": None,
-                "priority": 100,
-                "enabled": True,
-                "allowed_tools": [],
-                "blocked_tools": [],
-                "can_delegate_to": [],
-                "sandbox_timeout": 30,
-                "sandbox_network": "allow",
-            }
-
-        raise HTTPException(404, f"Agent '{agent_name}' not found")
-
-    @app.post("/api/v1/agents", dependencies=deps)
-    async def create_agent(request: Request) -> dict[str, Any]:
-        """Create a new agent profile."""
-        body = await request.json()
-        name = body.get("name", "").strip().lower().replace(" ", "-")
-        if not name:
-            raise HTTPException(400, "Name is required")
-
-        agents_path = config_manager.config.cognithor_home / "agents.yaml"
-        raw = {}
-        if agents_path.exists():
-            raw = yaml.safe_load(agents_path.read_text(encoding="utf-8")) or {}
-        agents = raw.get("agents", [])
-
-        # Check duplicate
-        if any(a.get("name") == name for a in agents):
-            raise HTTPException(409, f"Agent '{name}' already exists")
-
-        agent = {
-            "name": name,
-            "display_name": body.get("display_name", name.title()),
-            "description": body.get("description", ""),
-            "system_prompt": body.get("system_prompt", ""),
-            "language": body.get("language", "en"),
-            "preferred_model": body.get("preferred_model", ""),
-            "temperature": body.get("temperature", 0.7),
-            "top_p": body.get("top_p"),
-            "priority": body.get("priority", 0),
-            "enabled": body.get("enabled", True),
-            "allowed_tools": body.get("allowed_tools") or [],
-            "blocked_tools": body.get("blocked_tools", []),
-            "can_delegate_to": body.get("can_delegate_to", []),
-            "sandbox_timeout": body.get("sandbox_timeout", 30),
-            "sandbox_network": body.get("sandbox_network", "allow"),
-        }
-        agents.append(agent)
-        raw["agents"] = agents
-        agents_path.write_text(
-            yaml.dump(raw, default_flow_style=False, allow_unicode=True), encoding="utf-8"
-        )
-        # Live-reload agent router
-        router = getattr(gateway, "_agent_router", None)
-        if router and hasattr(router, "reload_from_yaml"):
-            router.reload_from_yaml(agents_path)
-        return {"status": "created", "agent": agent}
-
-    @app.put("/api/v1/agents/{agent_name}", dependencies=deps)
-    async def update_agent(agent_name: str, request: Request) -> dict[str, Any]:
-        """Update an existing agent profile."""
-        body = await request.json()
-        agents_path = config_manager.config.cognithor_home / "agents.yaml"
-        if not agents_path.exists():
-            raise HTTPException(404, f"Agent '{agent_name}' not found")
-
-        raw = yaml.safe_load(agents_path.read_text(encoding="utf-8")) or {}
-        agents = raw.get("agents", [])
-        found = False
-        updated_agent = None
-        for i, a in enumerate(agents):
-            if a.get("name") == agent_name:
-                # Update fields (including name rename)
-                for key in [
-                    "name",
-                    "display_name",
-                    "description",
-                    "system_prompt",
-                    "language",
-                    "preferred_model",
-                    "temperature",
-                    "top_p",
-                    "priority",
-                    "enabled",
-                    "allowed_tools",
-                    "blocked_tools",
-                    "can_delegate_to",
-                    "sandbox_timeout",
-                    "sandbox_network",
-                ]:
-                    if key in body:
-                        agents[i][key] = body[key]
-                found = True
-                updated_agent = agents[i]
-                break
-        if not found:
-            raise HTTPException(404, f"Agent '{agent_name}' not found")
-
-        raw["agents"] = agents
-        agents_path.write_text(
-            yaml.dump(raw, default_flow_style=False, allow_unicode=True), encoding="utf-8"
-        )
-        # Live-reload agent router
-        router = getattr(gateway, "_agent_router", None)
-        if router and hasattr(router, "reload_from_yaml"):
-            router.reload_from_yaml(agents_path)
-        return {"status": "updated", "agent": updated_agent}
-
-    @app.delete("/api/v1/agents/{agent_name}", dependencies=deps)
-    async def delete_agent(agent_name: str) -> dict[str, Any]:
-        """Delete an agent profile."""
-        if agent_name == "jarvis":
-            raise HTTPException(403, "Cannot delete the default agent")
-
-        agents_path = config_manager.config.cognithor_home / "agents.yaml"
-        if not agents_path.exists():
-            raise HTTPException(404, f"Agent '{agent_name}' not found")
-
-        raw = yaml.safe_load(agents_path.read_text(encoding="utf-8")) or {}
-        agents = raw.get("agents", [])
-        original_len = len(agents)
-        agents = [a for a in agents if a.get("name") != agent_name]
-        if len(agents) == original_len:
-            raise HTTPException(404, f"Agent '{agent_name}' not found")
-
-        raw["agents"] = agents
-        agents_path.write_text(
-            yaml.dump(raw, default_flow_style=False, allow_unicode=True), encoding="utf-8"
-        )
-        # Live-reload agent router
-        router = getattr(gateway, "_agent_router", None)
-        if router and hasattr(router, "reload_from_yaml"):
-            router.reload_from_yaml(agents_path)
-        return {"status": "deleted", "name": agent_name}
-
-    # -- Credentials ------------------------------------------------------
-
-    @app.get("/api/v1/credentials", dependencies=deps)
-    async def list_credentials() -> dict[str, Any]:
-        """Listet alle gespeicherten Credentials (nur Keys, keine Werte)."""
-        try:
-            from cognithor.security.credentials import CredentialStore
-
-            store = CredentialStore()
-            global_creds = store.list_entries()
-            return {
-                "credentials": [
-                    {"service": s, "key": k, "scope": "global"} for s, k in global_creds
-                ],
-            }
-        except Exception as exc:
-            log.error("credentials_list_failed", error=str(exc))
-            return {"credentials": [], "error": "Credentials konnten nicht geladen werden"}
-
-    @app.post("/api/v1/credentials", dependencies=deps)
-    async def store_credential(
-        request: Request,
-    ) -> dict[str, Any]:
-        """Speichert ein Credential (Body: service, key, value, agent_id)."""
-        body = await request.json()
-        service = body.get("service", "")
-        key = body.get("key", "")
-        value = body.get("value", "")
-        agent_id = body.get("agent_id", "")
-        if not service or not key or not value:
-            return {"error": "service, key und value sind erforderlich", "status": 400}
-        try:
-            from cognithor.security.credentials import CredentialStore
-
-            store = CredentialStore()
-            store.store(service, key, value, agent_id=agent_id)
-            return {"status": "ok", "service": service, "key": key, "scope": agent_id or "global"}
-        except Exception as exc:
-            log.error("credential_store_failed", error=str(exc))
-            return {"error": "Credential konnte nicht gespeichert werden", "status": 500}
-
-    @app.delete("/api/v1/credentials/{service}/{key}", dependencies=deps)
-    async def delete_credential(service: str, key: str, agent_id: str = "") -> dict[str, Any]:
-        """Loescht ein Credential."""
-        try:
-            from cognithor.security.credentials import CredentialStore
-
-            store = CredentialStore()
-            store.store(service, key, "", agent_id=agent_id)
-            return {"status": "ok", "deleted": f"{service}:{key}"}
-        except Exception as exc:
-            log.error("credential_delete_failed", error=str(exc))
-            return {"error": "Credential konnte nicht geloescht werden", "status": 500}
-
-    # -- Bindings ---------------------------------------------------------
-
-    @app.get("/api/v1/bindings", dependencies=deps)
-    async def list_bindings() -> dict[str, Any]:
-        """Listet alle Binding-Regeln aus bindings.yaml."""
-        try:
-            bindings_path = config_manager.config.cognithor_home / "bindings.yaml"
-            if bindings_path.exists():
-                raw = yaml.safe_load(bindings_path.read_text(encoding="utf-8")) or {}
-                bindings = raw.get("bindings", [])
-            else:
-                bindings = []
-            return {"bindings": bindings}
-        except Exception as exc:
-            log.error("bindings_list_failed", error=str(exc))
-            return {"bindings": [], "error": "Bindings konnten nicht geladen werden"}
-
-    @app.post("/api/v1/bindings", dependencies=deps)
-    async def create_binding(data: dict[str, Any]) -> dict[str, Any]:
-        """Erstellt oder aktualisiert eine Binding-Regel."""
-        try:
-            from cognithor.gateway.config_api import BindingRuleDTO
-            from cognithor.gateway.config_api import ConfigManager as CfgMgr
-
-            cfg_mgr = CfgMgr(config_manager.config)
-            dto = BindingRuleDTO(**data)
-            return {"binding": cfg_mgr.upsert_binding(dto), "status": "ok"}
-        except Exception as exc:
-            log.error("binding_create_failed", error=str(exc))
-            return {"error": "Binding konnte nicht erstellt werden", "status": 400}
-
-    @app.delete("/api/v1/bindings/{name}", dependencies=deps)
-    async def delete_binding(name: str) -> dict[str, Any]:
-        """Loescht eine Binding-Regel."""
-        try:
-            from cognithor.gateway.config_api import ConfigManager as CfgMgr
-
-            cfg_mgr = CfgMgr(config_manager.config)
-            if cfg_mgr.delete_binding(name):
-                return {"status": "ok", "deleted": name}
-            return {"error": f"Binding '{name}' nicht gefunden", "status": 404}
-        except Exception as exc:
-            log.error("binding_delete_failed", error=str(exc))
-            return {"error": "Binding konnte nicht geloescht werden", "status": 500}
-
-    # -- Circles ----------------------------------------------------------
-
-    @app.get("/api/v1/circles", dependencies=deps)
-    async def list_circles(peer_id: str = "") -> dict[str, Any]:
-        """Listet Trusted Circles."""
-        try:
-            from cognithor.skills.circles import CircleManager
-
-            circles_mgr = CircleManager()
-            circles = circles_mgr.list_circles(peer_id=peer_id)
-            return {
-                "circles": [
-                    {
-                        "circle_id": c.circle_id,
-                        "name": c.name,
-                        "description": c.description,
-                        "member_count": c.member_count,
-                        "curated_skills": len(c.curated_skills),
-                        "approved_skills": len(c.approved_skills()),
-                    }
-                    for c in circles
-                ],
-                "stats": circles_mgr.stats(),
-            }
-        except Exception as exc:
-            log.error("circles_list_failed", error=str(exc))
-            return {"circles": [], "error": "Circles konnten nicht geladen werden"}
-
-    @app.get("/api/v1/circles/stats", dependencies=deps)
-    async def circles_stats() -> dict[str, Any]:
-        """Ecosystem-Statistiken."""
-        try:
-            from cognithor.skills.circles import CircleManager
-
-            return CircleManager().stats()
-        except Exception as exc:
-            log.error("circles_stats_failed", error=str(exc))
-            return {"error": "Circle-Statistiken nicht verfuegbar"}
-
-    # -- Sandbox ----------------------------------------------------------
-
-    @app.get("/api/v1/sandbox", dependencies=deps)
-    async def get_sandbox() -> dict[str, Any]:
-        """Liest Sandbox-Konfiguration."""
-        try:
-            from cognithor.gateway.config_api import ConfigManager as CfgMgr
-
-            cfg_mgr = CfgMgr(config_manager.config)
-            return {"sandbox": cfg_mgr.get_sandbox()}
-        except Exception as exc:
-            log.error("sandbox_get_failed", error=str(exc))
-            return {"error": "Sandbox-Konfiguration nicht verfuegbar"}
-
-    @app.patch("/api/v1/sandbox", dependencies=deps)
-    async def update_sandbox(values: dict[str, Any]) -> dict[str, Any]:
-        """Aktualisiert Sandbox-Einstellungen."""
-        try:
-            from cognithor.gateway.config_api import ConfigManager as CfgMgr
-            from cognithor.gateway.config_api import SandboxUpdate
-
-            cfg_mgr = CfgMgr(config_manager.config)
-            update = SandboxUpdate(**values)
-            return {"sandbox": cfg_mgr.update_sandbox(update), "status": "ok"}
-        except Exception as exc:
-            log.error("sandbox_update_failed", error=str(exc))
-            return {"error": "Sandbox konnte nicht aktualisiert werden", "status": 400}
-
-    # -- Wizards ----------------------------------------------------------
-
-    @app.get("/api/v1/wizards", dependencies=deps)
-    async def list_wizards() -> dict[str, Any]:
-        """Alle verfuegbaren Konfigurations-Assistenten."""
-        from cognithor.gateway.wizards import WizardRegistry
-
-        reg = WizardRegistry()
-        return {"wizards": reg.list_wizards(), "count": reg.wizard_count}
-
-    @app.get("/api/v1/wizards/{wizard_type}", dependencies=deps)
-    async def get_wizard(wizard_type: str) -> dict[str, Any]:
-        """Details eines Wizards (Schritte + Templates)."""
-        from cognithor.gateway.wizards import WizardRegistry
-
-        reg = WizardRegistry()
-        wizard = reg.get(wizard_type)
-        if not wizard:
-            return {"error": f"Wizard '{wizard_type}' nicht gefunden"}
-        return wizard.to_dict()
-
-    @app.post("/api/v1/wizards/{wizard_type}/run", dependencies=deps)
-    async def run_wizard(wizard_type: str, body: dict[str, Any]) -> dict[str, Any]:
-        """Fuehrt einen Wizard aus und generiert Konfiguration."""
-        from cognithor.gateway.wizards import WizardRegistry
-
-        reg = WizardRegistry()
-        result = reg.run_wizard(wizard_type, body.get("values", {}))
-        if not result:
-            return {"error": f"Wizard '{wizard_type}' nicht gefunden"}
-        return result.to_dict()
-
-    @app.get("/api/v1/wizards/{wizard_type}/templates", dependencies=deps)
-    async def wizard_templates(wizard_type: str) -> dict[str, Any]:
-        """Templates eines Wizards."""
-        from cognithor.gateway.wizards import WizardRegistry
-
-        reg = WizardRegistry()
-        wizard = reg.get(wizard_type)
-        if not wizard:
-            return {"error": f"Wizard '{wizard_type}' nicht gefunden"}
-        return {
-            "templates": [
-                {
-                    "id": t.template_id,
-                    "name": t.name,
-                    "description": t.description,
-                    "icon": t.icon,
-                    "preset_values": t.preset_values,
-                }
-                for t in wizard.templates
-            ]
-        }
-
-    # -- RBAC -------------------------------------------------------------
-
-    @app.get("/api/v1/rbac/roles", dependencies=deps)
-    async def rbac_roles() -> dict[str, Any]:
-        """Alle verfuegbaren Rollen und ihre Berechtigungen."""
-        from cognithor.gateway.wizards import ROLE_PERMISSIONS
-
-        return {
-            "roles": {
-                role.value: {"permissions": [p.key for p in perms], "count": len(perms)}
-                for role, perms in ROLE_PERMISSIONS.items()
-            }
-        }
-
-    @app.get("/api/v1/rbac/check", dependencies=deps)
-    async def rbac_check(user_id: str, resource: str, action: str) -> dict[str, Any]:
-        """Prueft eine Berechtigung."""
-        from cognithor.gateway.wizards import RBACManager
-
-        mgr = RBACManager()
-        return {
-            "user_id": user_id,
-            "resource": resource,
-            "action": action,
-            "allowed": mgr.check_permission(user_id, resource, action),
-        }
-
-    # -- Auth Gateway -----------------------------------------------------
-
-    @app.get("/api/v1/auth/stats", dependencies=deps)
-    async def auth_stats() -> dict[str, Any]:
-        """Auth-Gateway-Statistiken."""
-        try:
-            from cognithor.gateway.auth import AuthGateway
-
-            return AuthGateway().stats()
-        except Exception as exc:
-            log.error("auth_stats_failed", error=str(exc))
-            return {"error": "Auth-Statistiken nicht verfuegbar"}
-
-    # -- Agent Heartbeat --------------------------------------------------
-
-    @app.get("/api/v1/agent-heartbeat/dashboard", dependencies=deps)
-    async def agent_heartbeat_dashboard() -> dict[str, Any]:
-        """Globale Dashboard-Uebersicht aller Agent-Heartbeats."""
-        try:
-            from cognithor.core.agent_heartbeat import AgentHeartbeatScheduler
-
-            return AgentHeartbeatScheduler().global_dashboard()
-        except Exception as exc:
-            log.error("heartbeat_dashboard_failed", error=str(exc))
-            return {"error": "Heartbeat-Dashboard nicht verfuegbar"}
-
-    @app.get("/api/v1/agent-heartbeat/{agent_id}", dependencies=deps)
-    async def agent_heartbeat_summary(agent_id: str) -> dict[str, Any]:
-        """Heartbeat-Zusammenfassung fuer einen Agent."""
-        try:
-            from cognithor.core.agent_heartbeat import AgentHeartbeatScheduler
-
-            return AgentHeartbeatScheduler().agent_summary(agent_id)
-        except Exception as exc:
-            log.error("heartbeat_summary_failed", agent_id=agent_id, error=str(exc))
-            return {"error": "Heartbeat-Zusammenfassung nicht verfuegbar"}
-
-
-# ======================================================================
-# Config read / write routes
-# ======================================================================
-
-
-def _register_config_routes(
-    app: Any,
-    deps: list[Any],
-    config_manager: ConfigManager,
-    gateway: Any = None,
-) -> None:
-    """Config CRUD, presets, reload."""
-
-    @app.get("/api/v1/health")
-    async def health_check() -> dict[str, Any]:
-        """Health check endpoint used by the Vite launcher."""
-        return {"status": "ok"}
-
-    @app.get("/api/v1/config", dependencies=deps)
-    async def get_config() -> dict[str, Any]:
-        """Gibt die gesamte Konfiguration zurueck (ohne Secrets)."""
-        data = config_manager.read()
-        data["_meta"] = {
-            "editable_sections": config_manager.editable_sections(),
-            "editable_top_level": config_manager.editable_top_level_fields(),
-        }
-        return data
-
-    @app.patch("/api/v1/config", dependencies=deps)
-    async def update_config_top_level(updates: dict[str, Any]) -> dict[str, Any]:
-        """Aktualisiert Top-Level-Felder."""
-        from cognithor.config_manager import _is_secret_field
-
-        results: list[dict[str, Any]] = []
-        for key, value in updates.items():
-            # Skip masked secret values — the UI sends "***" for untouched secrets.
-            # Real changes (new value or "") are passed through and persisted.
-            if value == "***" and _is_secret_field(key):
-                results.append({"key": key, "status": "skipped"})
-                continue
-            try:
-                config_manager.update_top_level(key, value)
-                results.append({"key": key, "status": "ok"})
-            except ValueError as exc:
-                log.error("config_update_key_failed", key=key, error=str(exc))
-                results.append({"key": key, "status": "error", "error": "Ungueltige Konfiguration"})
-        config_manager.save()
-        # Trigger live-reload of runtime components.
-        # Language changes must also reload the planner's cached system prompts
-        # (issue #136) — otherwise the planner keeps the German preset it
-        # loaded at startup and the LLM keeps answering in German.
-        if gateway is not None and hasattr(gateway, "reload_components"):
-            gateway.reload_components(
-                config=True,
-                prompts="language" in updates,
-            )
-        # Sync backend i18n locale when language changes (#33)
-        if "language" in updates:
-            try:
-                from cognithor.i18n import set_locale
-
-                set_locale(updates["language"])
-            except Exception:
-                log.debug("config_locale_sync_failed", exc_info=True)
-        return {"results": results}
-
-    @app.post("/api/v1/config/reload", dependencies=deps)
-    async def reload_config() -> dict[str, Any]:
-        """Laedt die Konfiguration neu aus der Datei."""
-        config_manager.reload()
-        if gateway is not None and hasattr(gateway, "reload_components"):
-            gateway.reload_components(prompts=True, policies=True, core_memory=True, config=True)
-        return {"status": "ok", "message": "Konfiguration und Komponenten neu geladen"}
-
-    @app.post("/api/v1/config/factory-reset", dependencies=deps)
-    async def factory_reset_config() -> dict[str, Any]:
-        """Reset configuration to defaults."""
-        from cognithor.config import CognithorConfig
-
-        try:
-            config_manager._config = CognithorConfig()
-            config_manager.save()
-            if gateway is not None and hasattr(gateway, "reload_components"):
-                gateway.reload_components(config=True)
-            return {"status": "ok", "message": "Configuration reset to defaults"}
-        except Exception as exc:
-            return {"error": str(exc)}
-
-    # -- Locales (available i18n language packs) --
-
-    @app.get("/api/v1/locales", dependencies=deps)
-    async def list_locales() -> dict[str, Any]:
-        """Returns available i18n locales and the currently active one."""
-        from cognithor.i18n import get_available_locales, get_locale
-        from cognithor.i18n.prompt_presets import available_preset_locales
-
-        locales = get_available_locales()
-        return {
-            "locales": locales,
-            "active": get_locale(),
-            "preset_locales": available_preset_locales(),
-        }
-
-    @app.post("/api/v1/translate-prompts", dependencies=deps)
-    async def translate_prompts(request: Request) -> dict[str, Any]:
-        """Translate system prompts to a target language.
-
-        Supports two methods:
-          - ``"preset"`` — Use curated prompt presets (instant, no LLM needed).
-          - ``"ollama"`` — Use local Ollama LLM to translate on-the-fly.
-
-        Request body::
-
-            {
-              "target_locale": "en",
-              "method": "preset" | "ollama",
-              "prompts": {                     // only for method=ollama
-                "plannerSystem": "...",
-                "replanPrompt": "...",
-                "escalationPrompt": "..."
-              }
-            }
-
-        Returns::
-
-            {
-              "translations": {
-                "plannerSystem": "...",
-                "replanPrompt": "...",
-                "escalationPrompt": "..."
-              },
-              "method": "preset" | "ollama"
-            }
-        """
-        try:
-            body = await request.json()
-        except Exception:
-            return {"error": "Invalid JSON body", "status": 400}
-
-        target_locale = body.get("target_locale", "").strip().lower()
-        method = body.get("method", "preset").strip().lower()
-
-        if not target_locale:
-            return {"error": "target_locale is required", "status": 400}
-        if method not in ("preset", "ollama"):
-            return {"error": "method must be 'preset' or 'ollama'", "status": 400}
-
-        # ── Method: preset ──────────────────────────────────────────
-        if method == "preset":
-            from cognithor.i18n.prompt_presets import get_preset
-
-            preset = get_preset(target_locale)
-            if preset is None:
-                from cognithor.i18n.prompt_presets import available_preset_locales
-
-                return {
-                    "error": f"No preset available for '{target_locale}'",
-                    "available": available_preset_locales(),
-                    "status": 404,
-                }
-            return {"translations": preset, "method": "preset"}
-
-        # ── Method: ollama ──────────────────────────────────────────
-        prompts = body.get("prompts", {})
-        if not prompts:
-            return {"error": "prompts dict is required for method=ollama", "status": 400}
-
-        cfg = config_manager.config
-        ollama_url = cfg.ollama.base_url.rstrip("/")
-        model = cfg.models.planner.name
-        timeout = cfg.ollama.timeout_seconds
-
-        import httpx
-
-        translations: dict[str, str] = {}
-        errors: list[str] = []
-
-        async with httpx.AsyncClient(timeout=timeout) as client:
-            for key, text in prompts.items():
-                if key not in ("plannerSystem", "replanPrompt", "escalationPrompt"):
-                    continue
-                if not text or not text.strip():
-                    continue
-
-                system_msg = (
-                    f"You are a professional translator. Translate the following system "
-                    f"prompt to {target_locale}. Preserve ALL template variables exactly "
-                    f"as they are (e.g., {{tools_section}}, {{owner_name}}, "
-                    f"{{results_section}}, {{original_goal}}, {{tool}}, {{reason}}). "
-                    f"Preserve all markdown formatting, code blocks, and JSON structures. "
-                    f"Output ONLY the translated text, nothing else."
-                )
-
-                try:
-                    resp = await client.post(
-                        f"{ollama_url}/api/chat",
-                        json={
-                            "model": model,
-                            "messages": [
-                                {"role": "system", "content": system_msg},
-                                {"role": "user", "content": text},
-                            ],
-                            "stream": False,
-                        },
-                    )
-                    resp.raise_for_status()
-                    data = resp.json()
-                    translated = data.get("message", {}).get("content", "").strip()
-
-                    # Validate that template variables survived translation
-                    import re
-
-                    original_vars = set(re.findall(r"\{[\w_]+\}", text))
-                    translated_vars = set(re.findall(r"\{[\w_]+\}", translated))
-                    missing = original_vars - translated_vars
-                    if missing:
-                        errors.append(f"{key}: template variables lost in translation: {missing}")
-                        # Still include the translation but flag it
-                        translations[key] = translated
-                    else:
-                        translations[key] = translated
-                except httpx.TimeoutException:
-                    errors.append(f"{key}: Ollama request timed out")
-                except httpx.HTTPStatusError as exc:
-                    errors.append(f"{key}: Ollama HTTP {exc.response.status_code}")
-                except Exception as exc:
-                    errors.append(f"{key}: {exc!s}")
-
-        result: dict[str, Any] = {"translations": translations, "method": "ollama"}
-        if errors:
-            result["warnings"] = errors
-        return result
-
-    # -- Presets (BEFORE {section} routes to avoid path parameter conflict) --
-
-    @app.get("/api/v1/config/presets", dependencies=deps)
-    async def list_presets() -> dict[str, Any]:
-        """Listet verfuegbare Konfigurations-Presets."""
-        return {
-            "presets": [
-                {
-                    "name": "minimal",
-                    "description": "Minimale Konfiguration (CLI-only, kleine Modelle)",
-                    "sections": {
-                        "channels": {
-                            "cli_enabled": True,
-                            "telegram_enabled": False,
-                            "webui_enabled": False,
-                        },
-                        "heartbeat": {"enabled": False},
-                        "dashboard": {"enabled": False},
-                    },
-                },
-                {
-                    "name": "standard",
-                    "description": "Standard-Setup (CLI + WebUI, Heartbeat, Dashboard)",
-                    "sections": {
-                        "channels": {"cli_enabled": True, "webui_enabled": True},
-                        "heartbeat": {"enabled": True, "interval_minutes": 30},
-                        "dashboard": {"enabled": True},
-                    },
-                },
-                {
-                    "name": "full",
-                    "description": "Vollausbau (alle Channels, Heartbeat, Dashboard, Plugins)",
-                    "sections": {
-                        "channels": {
-                            "cli_enabled": True,
-                            "webui_enabled": True,
-                            "telegram_enabled": True,
-                            "slack_enabled": True,
-                            "discord_enabled": True,
-                        },
-                        "heartbeat": {"enabled": True, "interval_minutes": 15},
-                        "dashboard": {"enabled": True},
-                        "plugins": {"auto_update": True},
-                    },
-                },
-            ],
-        }
-
-    @app.post("/api/v1/config/presets/{preset_name}", dependencies=deps)
-    async def apply_preset(preset_name: str) -> dict[str, Any]:
-        """Wendet ein Konfigurations-Preset an."""
-        presets = (await list_presets())["presets"]
-        preset = next((p for p in presets if p["name"] == preset_name), None)
-        if not preset:
-            return {"error": f"Preset '{preset_name}' nicht gefunden", "status": 404}
-        results = []
-        for section, values in preset["sections"].items():
-            try:
-                config_manager.update_section(section, values)
-                results.append({"section": section, "status": "ok"})
-            except ValueError as exc:
-                log.warning("preset_section_update_failed", section=section, error=str(exc))
-                results.append(
-                    {"section": section, "status": "error", "error": "Ungueltige Konfiguration"}
-                )
-        config_manager.save()
-        return {"preset": preset_name, "results": results}
-
-    # -- Config Versioning / Rollback ------------------------------------------
-
-    @app.get("/api/v1/config/revisions", dependencies=deps)
-    async def list_config_revisions() -> dict[str, Any]:
-        """List all saved config revisions (newest first)."""
-        from cognithor.core.config_versioning import list_revisions
-
-        return {"revisions": list_revisions()}
-
-    @app.post("/api/v1/config/rollback/{revision_id}", dependencies=deps)
-    async def rollback_config(revision_id: str) -> dict[str, Any]:
-        """Roll back the config to a previous revision.
-
-        Saves the current config as a new revision first, then applies
-        the historic config and persists it.
-        """
-        from cognithor.core.config_versioning import rollback_to, save_config_revision
-
-        # Save current state before rollback
-        try:
-            current = config_manager.read(include_secrets=True)
-            save_config_revision(current, reason=f"pre-rollback to {revision_id}")
-        except Exception:
-            log.warning("config_pre_rollback_save_failed", exc_info=True)
-
-        try:
-            historic_config = rollback_to(revision_id)
-        except FileNotFoundError as exc:
-            raise HTTPException(
-                status_code=404,
-                detail=f"Revision '{revision_id}' not found",
-            ) from exc
-        except ValueError as exc:
-            raise HTTPException(status_code=400, detail=str(exc)) from exc
-
-        # Apply the historic config via Pydantic validation
-        from pydantic import ValidationError
-
-        from cognithor.config import CognithorConfig
-
-        try:
-            new_cfg = CognithorConfig(**historic_config)
-        except ValidationError as exc:
-            raise HTTPException(
-                status_code=400,
-                detail=f"Historic config fails validation: {exc}",
-            ) from exc
-
-        config_manager._config = new_cfg
-        config_manager.save()
-
-        if gateway is not None and hasattr(gateway, "reload_components"):
-            gateway.reload_components(config=True)
-
-        return {"status": "ok", "revision_id": revision_id, "message": "Rollback applied"}
-
-    # -- Network Endpoints -------------------------------------------------------
-
-    @app.get("/api/v1/network/interfaces", dependencies=deps)
-    async def list_network_interfaces() -> dict[str, Any]:
-        """Detected network interfaces with enable/disable status."""
-        try:
-            from cognithor.core.network_endpoints import NetworkEndpointManager
-
-            mgr = NetworkEndpointManager()
-            return {"interfaces": mgr.get_detected_interfaces()}
-        except Exception as exc:
-            return {"interfaces": [], "error": str(exc)}
-
-    @app.put("/api/v1/network/endpoints", dependencies=deps)
-    async def update_network_endpoints(request: Request) -> dict[str, Any]:
-        """Update which network interfaces are enabled for API binding."""
-        body = await request.json()
-        try:
-            from cognithor.core.network_endpoints import NetworkEndpointManager
-
-            mgr = NetworkEndpointManager()
-            if "enabled_ips" in body:
-                mgr.set_enabled_ips(body["enabled_ips"])
-            if "auto_detect" in body:
-                mgr.set_auto_detect(body["auto_detect"])
-            return {
-                "status": "ok",
-                "bind_host": mgr.get_bind_host(),
-                "active_ips": mgr.get_active_ips(),
-                "message": "Restart required for bind changes to take effect",
-            }
-        except Exception as exc:
-            return {"error": str(exc)}
-
-    # -- Device Pairing (Mobile) ------------------------------------------------
-
-    @app.get("/api/v1/devices", dependencies=deps)
-    async def list_paired_devices() -> dict[str, Any]:
-        """List all paired devices."""
-        try:
-            from cognithor.security.device_pairing import DevicePairingManager
-
-            _secret = getattr(gateway, "_internal_api_token", "") if gateway else ""
-            mgr = DevicePairingManager(master_secret=_secret or "fallback")
-            return {"devices": mgr.list_devices()}
-        except Exception as exc:
-            return {"devices": [], "error": str(exc)}
-
-    @app.post("/api/v1/devices/pair", dependencies=deps)
-    async def pair_device(request: Request) -> dict[str, Any]:
-        """Create a new pairing token for a mobile device."""
-        body = await request.json()
-        device_name = body.get("name", "Unknown Device")
-        try:
-            from cognithor.security.device_pairing import DevicePairingManager
-
-            _secret = getattr(gateway, "_internal_api_token", "") if gateway else ""
-            mgr = DevicePairingManager(master_secret=_secret or "fallback")
-            pt = mgr.create_pairing_token(device_name)
-            _port = request.url.port or 8741
-            try:
-                from cognithor.utils.network import get_reachable_url
-
-                _reach_url = get_reachable_url("0.0.0.0", _port)
-                # Extract host from URL
-                from urllib.parse import urlparse
-
-                _host = urlparse(_reach_url).hostname or "127.0.0.1"
-            except ImportError:
-                _host = request.client.host if request.client else "127.0.0.1"
-            qr = mgr.qr_payload(pt, _host, _port)
-            return {
-                "device_id": pt.device_id,
-                "token": pt.token,
-                "expires_at": pt.expires_at,
-                "qr_payload": qr,
-            }
-        except Exception as exc:
-            return {"error": str(exc)}
-
-    @app.delete("/api/v1/devices/{device_id}", dependencies=deps)
-    async def revoke_device(device_id: str) -> dict[str, Any]:
-        """Revoke a paired device."""
-        try:
-            from cognithor.security.device_pairing import DevicePairingManager
-
-            _secret = getattr(gateway, "_internal_api_token", "") if gateway else ""
-            mgr = DevicePairingManager(master_secret=_secret or "fallback")
-            if mgr.revoke_device(device_id):
-                return {"status": "revoked", "device_id": device_id}
-            return {"error": "Device not found", "status": 404}
-        except Exception as exc:
-            return {"error": str(exc)}
-
-    # -- Config Section CRUD (AFTER presets to avoid {section} capturing "presets") --
-
-    @app.get("/api/v1/config/{section}", dependencies=deps)
-    async def get_config_section(section: str) -> dict[str, Any]:
-        """Gibt eine einzelne Konfigurations-Sektion zurueck."""
-        result = config_manager.read_section(section)
-        if result is None:
-            return {"error": f"Sektion '{section}' nicht gefunden", "status": 404}
-        return {"section": section, "values": result}
-
-    @app.patch("/api/v1/config/{section}", dependencies=deps)
-    async def update_config_section(section: str, values: dict[str, Any]) -> dict[str, Any]:
-        """Aktualisiert eine Konfigurations-Sektion."""
-        from cognithor.config_manager import _is_secret_field
-
-        def _deep_clean_secrets(
-            data: dict[str, Any],
-            existing: dict[str, Any] | None = None,
-            *,
-            _depth: int = 0,
-        ) -> dict[str, Any]:
-            """Recursively strip masked ('***') and empty secret values."""
-            if _depth > 5:
-                return data
-            out: dict[str, Any] = {}
-            for k, v in data.items():
-                if isinstance(v, dict):
-                    ex_sub = existing.get(k) if isinstance(existing, dict) else None
-                    cleaned_sub = _deep_clean_secrets(
-                        v,
-                        ex_sub if isinstance(ex_sub, dict) else None,
-                        _depth=_depth + 1,
-                    )
-                    if cleaned_sub:  # only include non-empty dicts
-                        out[k] = cleaned_sub
-                elif _is_secret_field(k):
-                    if v == "***":
-                        continue  # skip masked placeholders
-                    if (v == "" or v is None) and existing:
-                        ex_val = existing.get(k, "") if isinstance(existing, dict) else ""
-                        if ex_val and ex_val != "":
-                            continue  # protect existing non-empty secret
-                    out[k] = v
-                else:
-                    out[k] = v
-            return out
-
-        # Get existing section values (raw, unmasked) for protection comparison
-        raw_cfg = config_manager.config.model_dump(mode="json")
-        existing_section = (
-            raw_cfg.get(section, {}) if isinstance(raw_cfg.get(section), dict) else {}
-        )
-        cleaned = _deep_clean_secrets(values, existing_section)
-        try:
-            config_manager.update_section(section, cleaned)
-            config_manager.save()
-            # Trigger live-reload of runtime components (executor, web tools, model router)
-            if gateway is not None and hasattr(gateway, "reload_components"):
-                gateway.reload_components(config=True)
-
-            # Validate model availability when saving 'models' section
-            model_warnings: list[str] = []
-            if section == "models":
-                router = getattr(gateway, "_model_router", None) if gateway else None
-                if router and router._available_models:
-                    available = router._available_models
-                    cfg_models = config_manager.config.models
-                    for role, model_cfg in [
-                        ("planner", cfg_models.planner),
-                        ("executor", cfg_models.executor),
-                        ("coder", cfg_models.coder),
-                        ("embedding", cfg_models.embedding),
-                    ]:
-                        if model_cfg.name and model_cfg.name not in available:
-                            model_warnings.append(
-                                f"Modell '{model_cfg.name}' ({role}) nicht in Ollama gefunden. "
-                                f"Bitte installieren: ollama pull {model_cfg.name}"
-                            )
-
-            result: dict[str, Any] = {
-                "status": "ok",
-                "section": section,
-                "updated_keys": list(cleaned.keys()),
-            }
-            if model_warnings:
-                result["warnings"] = model_warnings
-            return result
-        except ValueError as exc:
-            log.warning("config_section_update_failed", section=section, error=str(exc))
-            return {"error": "Ungueltige Konfiguration", "status": 400}
 
 
 # ======================================================================
